requirements-local.txt
# Requirements.txt for local development on a Mac/Windows rather than the intended RPi Linux Device

# AudioHandler
sounddevice
wavio
pydub

# TtsHandler
ChatTTS

# ASR
transformers
numpy
torch

# Translation
sentencepiece

# IoT Node to Node Communication
paho-mqtt